// src/app/home/types/user.type.ts
export interface User {
    firstName: string;
    lastName: string;
    address: string;
    city: string;
    state: string;
    pin: string;
    email: string;
    password: string;
}

export interface UserLogin {
    email: string,
    password: string
}

export interface loginToken {
    token: string,
    expiresInSeconds: number
}